ce23bcf498d2284ed389e7b229e63aeaa0c96e91 gui: add path selection to recovery panel (edouard)
a42eb6d36a3501766dfdc981ea2346aa0c8394e8 gui: change wording coins view (edouard)
13248dbd026ef85bee147adbb0c6213c6fd017ec installer: add multipath support (edouard)
35dbb47bc1212b07859438fe989d4598e7eeb3e9 gui: update liana with multipath support (edouard)

Pull request description:

ACKs for top commit:
  edouardparis:
    Self-ACK ce23bcf498d2284ed389e7b229e63aeaa0c96e91

Tree-SHA512: 9a126262d403826c818aedeba919b9e222df82de204cf9ddbc3c928b6e4756308ea3c8b67cd25e1f86e6c283eb2edd54b48779199903660229e97bf40ded8af2

+++ gui/src/app/state/coins.rs
@@ -15,11 +15,11 @@ pub struct CoinsPanel {
     selected: Vec<usize>,
     warning: Option<Error>,
     /// timelock value to pass for the heir to consume a coin.
-    timelock: u32,
+    timelock: u16,
 }
 
 impl CoinsPanel {
-    pub fn new(coins: &[Coin], timelock: u32) -> Self {
+    pub fn new(coins: &[Coin], timelock: u16) -> Self {
         let mut panel = Self {
             coins: Vec::new(),
             selected: Vec::new(),
